Added Validation to the SettlementService also restructured DTO package.
